se crean el navbar, se agrega cierre de sesión. Se agrega vistas especificar para abogados y administradores en clientes, funcionalidad de clientes completada

+++ clients/views.py
@@ -2,11 +2,19 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from .models import Cliente
 from .forms import ClienteForm
+from django.contrib import messages
 
 # Vista para listar clientes
 @login_required
 def listar_clientes(request):
-    clientes = Cliente.objects.all()
+    # Verificar si el usuario tiene el rol de "moderador"
+    if request.user.groups.filter(name='Moderador').exists():
+        # Si es moderador, obtener todos los clientes
+        clientes = Cliente.objects.all()
+    else:
+        # Si es abogado, obtener solo los clientes asociados a él
+        clientes = Cliente.objects.filter(abogado=request.user)
+
     return render(request, 'listar_clientes.html', {'clientes': clientes})
 
 # Vista para crear un cliente
@@ -15,8 +23,10 @@ def crear_cliente(request):
     if request.method == 'POST':
         form = ClienteForm(request.POST)
         if form.is_valid():
-            form.save()
-            return redirect('list_clients')
+            cliente = form.save(commit=False)  # No guardar aún
+            cliente.abogado = request.user  # Asignar al abogado (usuario actual)
+            cliente.save()  # Ahora guardar el cliente
+            return redirect('clients:list_clients')
     else:
         form = ClienteForm()
     return render(request, 'crear_cliente.html', {'form': form})
@@ -29,7 +39,7 @@ def actualizar_cliente(request, pk):
         form = ClienteForm(request.POST, instance=cliente)
         if form.is_valid():
             form.save()
-            return redirect('list_clients') ## TO DO vista de actualizar cliente. 
+            return redirect('clients:list_clients') 
     else:
         form = ClienteForm(instance=cliente)
     return render(request, 'update_client.html', {'form': form, 'actualizar': True})
@@ -38,5 +48,10 @@ def actualizar_cliente(request, pk):
 @login_required
 def borrar_cliente(request, pk):
     cliente = get_object_or_404(Cliente, pk=pk)
-    cliente.delete()
-    return redirect('list_clients')
+
+    if request.user.groups.filter(name='Moderador').exists():
+        cliente.delete()
+        return redirect('clients:list_clients')
+    else:
+        messages.error(request, 'No tienes permiso para eliminar clientes. Comunicate con un administrador del sistema')
+        return redirect('clients:list_clients')
